Align MBORecord field types with official schema

+++ include/trading/marketdata/databento/MBORecord.hpp
@@ -6,8 +6,17 @@
 namespace CMETradingSystem::MarketData::Databento
 {
 
-// Databento Market By Order intermediate structure.
-// Binary DBN data will be decoded into this structure.
+// ============================================================
+// Databento Market By Order intermediate record
+// ============================================================
+//
+// Binary MBO data is decoded into this structure before being
+// copied into the generic DBNRecord representation.
+//
+// Types match the official Databento MBO schema.
+// Symbol is not physically stored in each MboMsg and therefore
+// remains empty until symbology resolution is added.
+// ============================================================
 
 struct MBORecord
 {
@@ -25,13 +34,13 @@ struct MBORecord
     int64_t price{0};
     uint32_t size{0};
 
-    uint16_t channel_id{0};
+    uint8_t channel_id{0};
     uint64_t order_id{0};
 
     uint8_t flags{0};
     int32_t ts_in_delta{0};
 
-    uint64_t sequence{0};
+    uint32_t sequence{0};
 
     std::string symbol;
 };
